requirements.txt
biopython==1.77
ipython==7.22.0
matplotlib==3.3.1
numpy==1.20.2
pandas==1.2.4
pyqt==5.9.2
